@@ -18,7 +18,7 @@
 	"bugs": "https://github.com/derhuerst/seebrucke-events-calendar-feed/issues",
 	"license": "ISC",
 	"engines": {
-		"node": ">=16"
+		"node": ">=24.0.0"
 	},
 	"dependencies": {
 		"connect": "^3.7.0",
