@@ -1,15 +1,10 @@
-// todo: use import assertions once they're supported by Node.js & ESLint
-// https://github.com/tc39/proposal-import-assertions
-import {createRequire} from 'module'
-const require = createRequire(import.meta.url)
-
 import createDebug from 'debug'
 import {DateTime} from 'luxon'
 import {deepStrictEqual} from 'assert'
 import fetch from 'node-fetch'
 import slug from 'slug'
 
-const pkg = require('../package.json')
+import pkg from '../package.json' with {type: 'json'}
 
 const DAY = 24 * 60 * 60 * 1000
 
