@@ -5,7 +5,7 @@ export default [
 	eslint.configs.recommended,
 	{
 		languageOptions: {
-			ecmaVersion: 2022,
+			ecmaVersion: 2025,
 			globals: globals.node,
 		},
 		rules: {
